% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.lambda.R
\name{calculate.lambda}
\alias{calculate.lambda}
\title{This function calculates the lambda necessary to achieve the desired number
of nonzero groups.}
\usage{
calculate.lambda(Xtu.vec, Xk.col.vec, nonzero.groups)
}
\arguments{
\item{Xtu.vec}{This contains a vector of l2 norms of Xtu}

\item{Xk.col.vec}{This contains the number of predictors for each group}

\item{nonzero.groups}{number of desired nonzero groups}
}
\description{
This function calculates the lambda necessary to achieve the desired number
of nonzero groups.
}
\examples{
calculate.lambda(Xtu.vec=temp$Xtu.vec,Xk.col.vec=temp$Xk.col.vec,nonzero.groups=nonzero.groups)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
