% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.shrinkage.binary.R
\name{calculate.shrinkage.binary}
\alias{calculate.shrinkage.binary}
\title{This function calculates which groups should be completely zeroed}
\usage{
calculate.shrinkage.binary(Xtu.list, lambda, alpha)
}
\arguments{
\item{Xtu.list}{list containing each group for vector v and index for each loading}

\item{lambda}{minimum lambda that achieves desired number of nonzero groups}

\item{alpha}{l1 sparseness penalty}
}
\description{
This function calculates which groups should be completely zeroed
}
\examples{
calculate.shrinkage.binary(Xtu.list=Xtu.list,lambda=lambda,alpha=alpha)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
