% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.sparse.lambda.R
\name{calculate.sparse.lambda}
\alias{calculate.sparse.lambda}
\title{This function calculates the lambda necessary to achieve the desired number
of nonzero groups while incorporating an l1 penalty on the vector}
\usage{
calculate.sparse.lambda(Xtu.list, upper.lambda, alpha, nonzero.groups)
}
\arguments{
\item{Xtu.list}{list containing each group for vector v and index for each loading}

\item{upper.lambda}{maximum lambda allowed for uniroot function}

\item{alpha}{l1 sparseness penalty}

\item{nonzero.groups}{desired number of nonzero groups remaining in vector v}
}
\description{
This function calculates the lambda necessary to achieve the desired number
of nonzero groups while incorporating an l1 penalty on the vector
}
\examples{
calculate.sparse.lambda(Xtu.list=Xtu.list, upper.lambda=upper.lambda, alpha=alpha, nonzero.groups=nonzero.groups)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
