% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.partition.Simpls.R
\name{cv.partition.Simpls}
\alias{cv.partition.Simpls}
\title{cross validation function for sparse partial least squares}
\usage{
cv.partition.Simpls(
  arg.sparse,
  df.partition,
  npc,
  n.folds,
  sparsity.type = c("loadings", "sumabs"),
  nonzero.loadings = NULL,
  sumabsv = NULL,
  kernel = c("linear", "delta"),
  niter,
  trace
)
}
\arguments{
\item{arg.sparse}{vector for folds and sparse parameter}

\item{df.partition}{list of folds}

\item{npc}{number of desired principal components}

\item{n.folds}{number of folds to perform cross validation}

\item{sparsity.type}{specifies mechanism for vector shrinkage}

\item{nonzero.loadings}{specifies the desired number of nonzero loadings}

\item{sumabsv}{shrinkage using sum of absolute value of loadings}

\item{kernel}{specification of the response kernel}

\item{niter}{number of iterations in the SMD algorithm}

\item{trace}{display algorithm progress}
}
\description{
cross validation function for sparse partial least squares
}
\examples{
cv.partiion.Simpls(arg.sparse=c(test.index,sparse.lambda), df.partition=groupdata2.partition,npc=2,n.folds=5,sparsity.type="sumabs",nonzero.loadings=NULL,sumabsv=sqrt(10),kernel="linear",niter=200,trace=F)
}
\keyword{Least}
\keyword{Partial}
\keyword{Sparse}
\keyword{Squares}
