% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.sgSIMPLS.R
\name{cv.sgSIMPLS}
\alias{cv.sgSIMPLS}
\title{cross validation function for groupwise sparse SIMPLS}
\usage{
cv.sgSIMPLS(
  X,
  Y,
  npc,
  n.folds = 5,
  kernel = c("linear", "delta"),
  groups = NULL,
  nonzero.groups = NULL,
  alpha = NULL,
  trace = F,
  niter = 200,
  parallel = F,
  n.cores = NULL,
  ind.names = NULL,
  part.balance = F,
  mc.method = T
)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{npc}{number of desired principal components}

\item{n.folds}{number of folds to perform cross validation}

\item{kernel}{specification of the response kernel}

\item{groups}{vector of groups to which each predictor belongs}

\item{nonzero.groups}{number of desired nonzero groups}

\item{alpha}{l1 penalty term}

\item{trace}{display algorithm progress}

\item{niter}{number of iterations in the SMD algorithm}

\item{parallel}{flag for parallel process to parApply}

\item{n.cores}{number of cores to be used in parallel process}

\item{ind.names}{vector of each observation label}

\item{part.balance}{flag for whether folds process should balance factors}

\item{mc.method}{flag for whether the parallel method should use mclapply}
}
\description{
cross validation function for groupwise sparse SIMPLS
}
\examples{
cv.sgSIMPLS <- function(X,Y,npc, n.folds=5, kernel=c("linear","delta"), groups=NULL, nonzero.groups=NULL, alpha=NULL, trace=F,niter=200,parallel=F,n.cores=NULL, ind.names=NULL, part.balance=F)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
