% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.SSPCA.R
\name{cv.SSPCA}
\alias{cv.SSPCA}
\title{cross validation function for sparse supervised PCA}
\usage{
cv.SSPCA(
  X,
  Y,
  npc,
  n.folds,
  sparsity.type = c("loadings", "sumabs"),
  nonzero.loadings = NULL,
  sumabsv = NULL,
  kernel = c("linear", "delta"),
  parallel = F,
  n.cores = NULL,
  niter = 50,
  trace = F,
  part.balance = T,
  mc.method = T
)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{npc}{number of desired principal components}

\item{n.folds}{number of folds to perform cross validation}

\item{sparsity.type}{specifies mechanism for vector shrinkage}

\item{nonzero.loadings}{specifies the desired number of nonzero loadings}

\item{sumabsv}{shrinkage using sum of absolute value of loadings}

\item{kernel}{specification of the response kernel}

\item{parallel}{flag for parallel process to parApply}

\item{n.cores}{number of cores to be used in parallel process}

\item{niter}{number of iterations in the SMD algorithm}

\item{trace}{display algorithm progress}

\item{part.balance}{flag for whether folds process should balance factors}

\item{mc.method}{flag for whether the parallel method should use mclapply}
}
\description{
cross validation function for sparse supervised PCA
}
\examples{
cv.SSPCA <- function(X=X, Y=Y ,npc=2, n.folds=5, sparsity.type="sumabs", nonzero.loadings=NULL, sumabsv=sqrt(10),  kernel="linear", parallel=F, n.cores=NULL, niter=200, trace=F, part.balance=F)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
