% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_partition_group.R
\name{cv_partition_group}
\alias{cv_partition_group}
\title{cross validation function for groupwise supervised PCA}
\usage{
cv_partition_group(
  arg.group,
  df.partition,
  npc,
  n.folds,
  groups,
  kernel = c("linear", "delta"),
  ind.names,
  niter,
  trace
)
}
\arguments{
\item{arg.group}{vector for folds and sparse parameter}

\item{df.partition}{list of folds}

\item{npc}{number of desired principal components}

\item{n.folds}{number of folds to perform cross validation}

\item{groups}{vector of groups to which each predictor belongs}

\item{kernel}{specification of the response kernel}

\item{ind.names}{vector of each observation label}

\item{niter}{number of iterations in the SMD algorithm}

\item{trace}{display algorithm progress}
}
\description{
cross validation function for groupwise supervised PCA
}
\examples{
cv_partition_group(arg.group=c(test.index,sparse.lambda), df.partition=groupdata2.partition,npc=2,n.folds=5,kernel="linear",ind.names=NULL,niter=200,trace=F)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
