% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflate.R
\name{deflate}
\alias{deflate}
\title{This function deflates the matrix X based on components V}
\usage{
deflate(X, V, n, p)
}
\arguments{
\item{X}{predictor matrix}

\item{V}{component vectors}

\item{n}{number of observations}

\item{p}{number of predictors}
}
\description{
This function deflates the matrix X based on components V
}
\examples{
deflate(X=X,V=V,n=200,p=1000)
}
\keyword{Sparse}
\keyword{least}
\keyword{partial}
\keyword{squares}
