% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.v.R
\name{embed.v}
\alias{embed.v}
\title{This function replaces the oldings of the old v vector with the new values}
\usage{
embed.v(shrunk.Xtu.list, vold)
}
\arguments{
\item{shrunk.Xtu.list}{list of shrunk loadings for each group with corresponding index}

\item{vold}{old v vector to be replaced with new loadings}
}
\description{
This function replaces the oldings of the old v vector with the new values
}
\examples{
embed.v(shrunk.Xtu.list=shrunk.list,vold=v)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
