% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_group.R
\name{extract_group}
\alias{extract_group}
\title{This extracts and returns the left and right singular vectors using shrinkage on the predictor vector by groups}
\usage{
extract_group(
  X,
  n,
  p,
  groups,
  nonzero.groups,
  unique.group,
  num.group,
  niter,
  trace,
  npc
)
}
\arguments{
\item{X}{predictor matrix}

\item{n}{number of observations}

\item{p}{number of predictors}

\item{groups}{vector containing groups for which each predictor belongs}

\item{nonzero.groups}{number of desired nonzero groups}

\item{unique.group}{vector containing unique group listings}

\item{num.group}{number of unique groups}

\item{niter}{number of iterations}

\item{trace}{whether to output progress}

\item{npc}{which pc is being extracted}
}
\description{
This extracts and returns the left and right singular vectors using shrinkage on the predictor vector by groups
}
\examples{
extract_group(X, n, p, groups, nonzero.groups, unique.group, num.group, niter, trace, npc)
}
\keyword{Sparse}
\keyword{elast}
\keyword{partial}
\keyword{squares}
