% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_load.R
\name{extract_load}
\alias{extract_load}
\title{This extracts and returns the left and right singular vectors}
\usage{
extract_load(X, n, p, nonzero.loadings, niter, trace, npc)
}
\arguments{
\item{X}{predictor matrix}

\item{n}{number of observations}

\item{p}{number of predictors}

\item{niter}{number of iterations}

\item{trace}{whether to output progress}

\item{npc}{which pc is being extracted}

\item{load}{value for number of desired nonzero loadings}
}
\description{
This extracts and returns the left and right singular vectors
}
\examples{
extract_load(X=X,n=200,p=1000,nonzero.loadings=10,niter=200,trace=F,npc=2)
}
\keyword{Sparse}
\keyword{elast}
\keyword{partial}
\keyword{squares}
