% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.psi.R
\name{get.psi}
\alias{get.psi}
\title{This function accepts a data matrix and the response kernel matrix
and generates an nxp matrix that has supervision}
\usage{
get.psi(X, L, n, p)
}
\arguments{
\item{X}{nxp matrix of predictors}

\item{L}{nxn response kernel}

\item{n}{an integer storing the number response observations}

\item{p}{an integer storing the number of predictors}
}
\description{
This function accepts a data matrix and the response kernel matrix
and generates an nxp matrix that has supervision
}
\examples{
get.psi(X=X,L=L,n=200,p=1000)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
