% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Xk.R
\name{get.Xk}
\alias{get.Xk}
\title{This function calculates the v vector loadings, l2 norms for each group, and number of preditors in each group}
\usage{
get.Xk(X, u, groups, unique.group, num.group)
}
\arguments{
\item{X}{predictor matrix}

\item{u}{iterated vector corresponding to observations}

\item{groups}{vector of groups to which each predictor belongs}

\item{unique.group}{list of unique group labesl}

\item{num.group}{number of unique groups}
}
\description{
This function calculates the v vector loadings, l2 norms for each group, and number of preditors in each group
}
\examples{
get.Xk(X=X,u=u,groups=groups,unique.group=unique.group,num.group=num.group)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
