% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsoft.R
\name{gsoft}
\alias{gsoft}
\title{This function performs group soft thresholding on a vector.}
\usage{
gsoft(pk, l2, lambda)
}
\arguments{
\item{pk}{kth group of vector}

\item{l2}{l2 norm}

\item{lambda}{shrinkage penalty term}
}
\description{
This function performs group soft thresholding on a vector.
}
\examples{
soft(x=v,d=lambda)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
