% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2n.R
\name{l2n}
\alias{l2n}
\title{This function calculates the l2 norm of a vector}
\usage{
l2n(vec)
}
\arguments{
\item{vec}{input vector}
}
\description{
This function calculates the l2 norm of a vector
}
\examples{
l2n(vec=xvec)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
