% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LamSearch.R
\name{LamSearch}
\alias{LamSearch}
\title{This function peforms identifies the desired lambda penalty resulting
in the desired number of nonzero loadings in the v matrix}
\usage{
LamSearch(v, nzl)
}
\arguments{
\item{v}{vector to undergo shrinkage}

\item{nzl}{number of desired nonzero loadings}
}
\description{
This function peforms identifies the desired lambda penalty resulting
in the desired number of nonzero loadings in the v matrix
}
\examples{
LamSearch(v=vec,nzl=10)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
