% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManPlot.R
\name{ManPlot}
\alias{ManPlot}
\title{This function provides a manhattan plot for the magnitude of eigenvector loadings.}
\usage{
ManPlot(
  V,
  predictor.names = NULL,
  group.structure = NULL,
  specific.groups = NULL,
  exclude.groups = NULL,
  man.thresh = 0,
  special.colors = NULL,
  optional.index = NULL,
  point.size = 1.5
)
}
\arguments{
\item{V}{contains the eigenevectors corresponding to the loadings}

\item{predictor.names}{names of predictors corresponding to loadings in V}

\item{group.structure}{grouping structure of loadings in the Manhattan Plot}

\item{specific.groups}{allows certain groups to be selected}

\item{exclude.groups}{allows certain groups to be excluded}

\item{man.thresh}{draws a horizontal line at a specified threshold}

\item{special.colors}{data frame of group structures with corresponding color palettes}

\item{optional.index}{gives values if index values differ from predictor order in V}

\item{point.size}{size of the plot points}
}
\description{
This function provides a manhattan plot for the magnitude of eigenvector loadings.
}
\examples{
CifManPlot(Zlist, sepAnalysis, nresp, man.thresh, resp.names, 
  specific.groups=NULL, exclude.groups=NULL, special.colors=NULL)
}
\keyword{CIFTI}
\keyword{manhattan}
\keyword{plot}
