% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.fill.R
\name{mat.fill}
\alias{mat.fill}
\title{cross validation function for sparse supervised PCA}
\usage{
mat.fill(param.grid, n.sp, n.folds)
}
\arguments{
\item{n.sp}{number of sparse parameter options}

\item{n.folds}{number of folds}
}
\description{
cross validation function for sparse supervised PCA
}
\examples{
mat.fill(param.grid=param.grid,n.sp=100,n.folds=5)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
