% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.fill.R
\name{matrix.fill}
\alias{matrix.fill}
\title{cross validation function for sparse supervised PCA}
\usage{
matrix.fill(param.grid, n.args, n.folds)
}
\arguments{
\item{n.args}{number of grid search parameter options}

\item{n.folds}{number of folds}
}
\description{
cross validation function for sparse supervised PCA
}
\examples{
matrix.fill(param.grid=param.grid,n.args=n.args,n.folds=n.folds)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
