% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.metric.Simpls.R
\name{model.metric.Simpls}
\alias{model.metric.Simpls}
\title{cross validation function for sparse supervised PCA}
\usage{
model.metric.Simpls(Y, X, V, mod, kernel = c("linear", "delta"))
}
\arguments{
\item{Y}{vector of response}

\item{X}{predictor matrix}

\item{V}{pca vector}

\item{mod}{holds the learned model}

\item{kernel}{flag for kernel type}
}
\description{
cross validation function for sparse supervised PCA
}
\examples{
model.metric.Simpls(Y=Y,X=X,V=V,mod=sspca.obj,kernel="linear")
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
