% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMD_group.R
\name{PMD_group}
\alias{PMD_group}
\title{This function peforms penalized matrix decomposition on a supplied matrix
and uses nonzero groups as the shrinkage mechanism}
\usage{
PMD_group(
  X,
  npc,
  n,
  p,
  groups,
  nonzero.groups,
  unique.group,
  num.group,
  niter,
  trace
)
}
\arguments{
\item{X}{matrix to undergo PMD}

\item{npc}{number of desired principal components}

\item{n}{number of observations}

\item{p}{number of predictors}

\item{groups}{vector of groups to which each predictor belongs}

\item{nonzero.groups}{number of desired nonzero groups}

\item{unique.group}{vector of each unique group label}

\item{num.group}{number of unique groups}

\item{niter}{number of iterations for the algorithm}

\item{trace}{displays information about progress of algorithm}
}
\description{
This function peforms penalized matrix decomposition on a supplied matrix
and uses nonzero groups as the shrinkage mechanism
}
\examples{
PMD_group(X=Psi,npc=npc,n=n,p=p,groups=groups,nonzero.groups=nonzero.groups,unique.group=unique.group,num.group=num.group,niter=niter,trace=trace)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
