% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMDvL1.load.R
\name{PMDvL1.load}
\alias{PMDvL1.load}
\title{This function peforms penalized matrix decomposition on a supplied matrix
and uses nonzero loadings as the shrinkage mechanism}
\usage{
PMDvL1.load(X, npc, n, p, nonzero.loadings, niter, trace)
}
\arguments{
\item{X}{matrix to undergo PMD}

\item{npc}{number of desired principal components}

\item{n}{number of observations}

\item{p}{number of predictors}

\item{nonzero.loadings}{desired number of nonzero loadings}

\item{niter}{number of iterations for the algorithm}

\item{trace}{displays information about progress of algorithm}
}
\description{
This function peforms penalized matrix decomposition on a supplied matrix
and uses nonzero loadings as the shrinkage mechanism
}
\examples{
PMDvL1.load(X=X,npc=2,n=200,p=1000,nonzero.loadings=10,niter=200,trace=F)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
