% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMDvL1.sumabs.R
\name{PMDvL1.sumabs}
\alias{PMDvL1.sumabs}
\title{This function peforms penalized matrix decomposition on a supplied matrix
and uses sum of absolute value of loadings as the shrinkage mechanism}
\usage{
PMDvL1.sumabs(X, npc, n, p, sumabsv, niter, trace)
}
\arguments{
\item{X}{matrix to undergo PMD}

\item{npc}{number of desired principal components}

\item{n}{number of observations}

\item{p}{number of predictors}

\item{niter}{number of iterations for the algorithm}

\item{trace}{displays information about progress of algorithm}

\item{sumasv}{sum of absolute value of loadings of the v vector}
}
\description{
This function peforms penalized matrix decomposition on a supplied matrix
and uses sum of absolute value of loadings as the shrinkage mechanism
}
\examples{
PMDvL1.sumabs(X=X,npc=2,n=200,p=1000,sumabsv=sqrt(10),niter=200,trace=F)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
