% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respkernel.R
\name{respkernel}
\alias{respkernel}
\title{This function accepts a response vector and converts into either a linear or
a delta kernel}
\usage{
respkernel(Y, n, kernel)
}
\arguments{
\item{Y}{initial response vector}

\item{n}{an integer storing the number response observations}

\item{kernel}{stores they type of kernel we wish to build, either linear or
delta}
}
\description{
This function accepts a response vector and converts into either a linear or
a delta kernel
}
\examples{
respkernel(Y=Y,n=200,kernel="linear")
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
