% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgv.R
\name{sgv}
\alias{sgv}
\title{Calculates shrinkage for sparse group vector v}
\usage{
sgv(lambda, v, alpha)
}
\arguments{
\item{lambda}{group shrinkage parameter}

\item{v}{vector to undergo shrinkage}

\item{alpha}{l1 shrinkage penalty term}
}
\description{
Calculates shrinkage for sparse group vector v
}
\examples{
soft(x=v,d=lambda)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
