% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD.load.R
\name{SMD.load}
\alias{SMD.load}
\title{This function peforms penalized matrix decomposition for one pair of
u and v vectors
Uses number of nonzero loadings for the shrinkage mechanism}
\usage{
SMD.load(X, d, u, v, n, p, nonzero.loadings, niter, trace, npc)
}
\arguments{
\item{X}{matrix to undergo PMD}

\item{d}{supplied singular value}

\item{u}{supplied left vector}

\item{v}{supplied right vector}

\item{n}{number of observations}

\item{p}{number of predictors}

\item{nonzero.loadings}{number of desired nonzero loadings}

\item{niter}{number of iterations for the algorithm}

\item{trace}{displays information about progress of algorithm}

\item{npc}{number of desired principal components}
}
\description{
This function peforms penalized matrix decomposition for one pair of
u and v vectors
Uses number of nonzero loadings for the shrinkage mechanism
}
\examples{
SMD.load(X=X,d=D,u=U,v=V,n=200,p=1000,nonzero.loadings=10,niter=200,trace=F,npc=1)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
