% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD_group.R
\name{SMD_group}
\alias{SMD_group}
\title{This function peforms penalized matrix decomposition for one pair of
u and v vectors
Uses number of nonzero groups for the shrinkage mechanism}
\usage{
SMD_group(
  X,
  d,
  u,
  v,
  n,
  p,
  groups,
  nonzero.groups,
  unique.group,
  num.group,
  niter,
  trace,
  npc
)
}
\arguments{
\item{X}{matrix to undergo PMD}

\item{d}{supplied singular value}

\item{u}{supplied left vector}

\item{v}{supplied right vector}

\item{n}{number of observations}

\item{p}{number of predictors}

\item{groups}{vector of groups to which each predictor belongs}

\item{nonzero.groups}{number of desired nonzero groups}

\item{unique.group}{vector of each unique group label}

\item{num.group}{number of unique groups}

\item{niter}{number of iterations for the algorithm}

\item{trace}{displays information about progress of algorithm}

\item{npc}{number of desired principal components}
}
\description{
This function peforms penalized matrix decomposition for one pair of
u and v vectors
Uses number of nonzero groups for the shrinkage mechanism
}
\examples{
SMD.group(X=Xuse, d=temp$d, u=temp$u, v=temp$v, n=n, p=p, groups=groups, nonzero.groups=nonzero.groups, unique.group=unique.group, num.group=num.group,niter=niter, trace=trace, npc=i)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
