% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD_sparse_group.R
\name{SMD_sparse_group}
\alias{SMD_sparse_group}
\title{This function peforms penalized matrix decomposition for one pair of
u and v vectors
Uses number of nonzero groups for the shrinkage mechanism as well as l1 penalty}
\usage{
SMD_sparse_group(
  X,
  d,
  u,
  v,
  n,
  p,
  groups,
  nonzero.groups,
  unique.group,
  num.group,
  alpha,
  niter,
  trace,
  npc
)
}
\arguments{
\item{X}{matrix to undergo PMD}

\item{d}{supplied singular value}

\item{u}{supplied left vector}

\item{v}{supplied right vector}

\item{n}{number of observations}

\item{p}{number of predictors}

\item{groups}{vector of groups to which each predictor belongs}

\item{nonzero.groups}{number of desired nonzero groups}

\item{unique.group}{vector of each unique group label}

\item{num.group}{number of unique groups}

\item{alpha}{l1 shrinkage penalty}

\item{niter}{number of iterations for the algorithm}

\item{trace}{displays information about progress of algorithm}

\item{npc}{number of desired principal components}
}
\description{
This function peforms penalized matrix decomposition for one pair of
u and v vectors
Uses number of nonzero groups for the shrinkage mechanism as well as l1 penalty
}
\examples{
SMD_sparse_group(X=Xuse, d=temp$d, u=temp$u, v=temp$v, n=n, p=p, groups=groups, nonzero.groups=nonzero.groups, unique.group=unique.group, alpha=alpha,num.group=num.group, niter=niter, trace=trace, npc=i)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
