% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMDorth.sumabs.R
\name{SMDorth.sumabs}
\alias{SMDorth.sumabs}
\title{This function peforms penalized matrix decomposition for one pair of
u and v vectors.  This uses orthogonilization for the previous U vectors.
Uses sum of absolute value of loadings as the shrinkage mechanism}
\usage{
SMDorth.sumabs(X, us, d, u, v, n, p, sumabsv, niter, trace, npc)
}
\arguments{
\item{X}{matrix to undergo PMD}

\item{d}{supplied singular value}

\item{u}{supplied left vector}

\item{v}{supplied right vector}

\item{n}{number of observations}

\item{p}{number of predictors}

\item{niter}{number of iterations for the algorithm}

\item{trace}{displays information about progress of algorithm}

\item{npc}{number of desired principal components}

\item{sumasv}{sum of absolute value of loadings of the v vector}
}
\description{
This function peforms penalized matrix decomposition for one pair of
u and v vectors.  This uses orthogonilization for the previous U vectors.
Uses sum of absolute value of loadings as the shrinkage mechanism
}
\examples{
SMDorth.sumabs(X=X,us=Us,d=D,u=U,v=V,n=200,p=1000,sumabsv=sqrt(10),niter=200,trace=F,npc=2)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
