% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.R
\name{soft}
\alias{soft}
\title{This function performs soft thresholding on a vector.}
\usage{
soft(x, d)
}
\arguments{
\item{x}{vector to undergo soft thresholding}

\item{d}{value for absolute value loading shrinkage}
}
\description{
This function performs soft thresholding on a vector.
}
\examples{
soft(x=v,d=lambda)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
