% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPCA.R
\name{SPCA}
\alias{SPCA}
\title{This function is the parent function for performing SPCA}
\usage{
SPCA(X, Y, npc, kernel = c("linear", "delta"))
}
\arguments{
\item{X}{nxp matrix of predictors}

\item{Y}{nx1 response vector}

\item{npc}{number of desired principal components}

\item{kernel}{designation for either linear or delta kernel}
}
\description{
This function is the parent function for performing SPCA
}
\examples{
SPCA(X=X,Y=Y,npc=2,kernel="linear")
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
