% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPCA.svd.R
\name{SPCA.svd}
\alias{SPCA.svd}
\title{This function performs supervised PCA on the psi matrix}
\usage{
SPCA.svd(X, npc)
}
\arguments{
\item{X}{nxp matrix to undergo supervised PCA}

\item{npc}{number of desired principal components}
}
\description{
This function performs supervised PCA on the psi matrix
}
\examples{
SPCA.svd(X=X,npc=2)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
