% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ssimpls.R
\name{Ssimpls}
\alias{Ssimpls}
\title{This function is the parent function for sparse partial least squares}
\usage{
Ssimpls(
  X,
  Y,
  npc,
  sparsity.type = c("loadings", "sumabs"),
  nonzero.loadings = NULL,
  sumabsv = NULL,
  kernel = c("linear", "delta"),
  niter = 50,
  trace = F
)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{npc}{number of desired principal components}

\item{sparsity.type}{specifies the type of sparsity constraint to be used}

\item{nonzero.loadings}{desired number of nonzero loadings}

\item{sumabsv}{shrinkage using sum of absolute value of loadings}

\item{kernel}{specification of the response kernel}

\item{niter}{number of iterations in the SMD algorithm}

\item{trace}{display algorithm progress}
}
\description{
This function is the parent function for sparse partial least squares
}
\examples{
Ssimpls(X=X,Y=Y,npc=2,sparsity.type="sumabs",nonzero.loadings=NULL,sumabsv=sqrt(10),kernel="linear",niter=200,trace=F)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
