% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSPCA.sumabs.R
\name{SSPCA.sumabs}
\alias{SSPCA.sumabs}
\title{This function performs sparse supervised PCA using sum of absolute values
of loadings for the v vector as the shrinkage mechanism}
\usage{
SSPCA.sumabs(
  X,
  Y,
  npc,
  sumabsv = NULL,
  kernel = c("linear", "delta"),
  niter = 50,
  trace = F
)
}
\arguments{
\item{X}{nxp predictor matrix}

\item{Y}{nx1 response vector}

\item{npc}{number of desired principal components}

\item{sumabsv}{shrinkage using sum of absolute value of loadings}

\item{kernel}{specification of the response kernel}

\item{niter}{number of iterations in the SMD algorithm}

\item{trace}{display algorithm progress}
}
\description{
This function performs sparse supervised PCA using sum of absolute values
of loadings for the v vector as the shrinkage mechanism
}
\examples{
SSPCA.sumabs(X=X,Y=Y,npc=2,sumabsv=sqrt(10),kernel="linear",niter=200,trace=F)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
