% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.shrinkage.R
\name{vector.shrinkage}
\alias{vector.shrinkage}
\title{This function shrinks each group of the v vector by the deisred lambda penalty term
to achieve the desired number of nonzero groups in vector v.}
\usage{
vector.shrinkage(Xtu.list, Xtu.vec, Xk.col.vec, lambda)
}
\arguments{
\item{Xtu.list}{list of estimated v vector groups with index for each loading}

\item{Xtu.vec}{This contains a vector of l2 norms of Xtu}

\item{Xk.col.vec}{This contains the number of predictors for each group}

\item{lambda}{group shrinkate penalty}
}
\description{
This function shrinks each group of the v vector by the deisred lambda penalty term
to achieve the desired number of nonzero groups in vector v.
}
\examples{
vector.shrinkage(Xtu.list = temp$Xtu.list, Xtu.vec=temp$Xtu.vec, Xk.col.vec=temp$Xk.col.vec, lambda=lambda)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
