% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.sparse.shrinkage.R
\name{vector.sparse.shrinkage}
\alias{vector.sparse.shrinkage}
\title{This function shrinks each group of the v vector by the deisred lambda penalty term
to achieve the desired number of nonzero groups in vector v while achieving l1 shrinkage.}
\usage{
vector.sparse.shrinkage(Xtu.list, shrink.list, lambda, alpha)
}
\arguments{
\item{Xtu.list}{list of estimated v vector groups with index for each loading}

\item{shrink.list}{list of groups to be zeroed out}

\item{lambda}{group shrinkate penalty}

\item{alpha}{l1 penalty term}
}
\description{
This function shrinks each group of the v vector by the deisred lambda penalty term
to achieve the desired number of nonzero groups in vector v while achieving l1 shrinkage.
}
\examples{
vector.sparse.shrinkage(Xtu.list=temp$Xtu.list, shrink.list=bin.list, lambda=lambda, alpha=alpha)
}
\keyword{Analysis}
\keyword{Component}
\keyword{Princpal}
\keyword{Supervised}
